refactor(ui): uicore の入れ子 internal をやめ depguard へ一本化する
widgets/internal/uicore を widgets/uicore へ移し、画面層からの遮断を
コンパイラでなく depguard の ui_core_inbound_guard で行う。

パスから internal が二重に現れるのを解消し、UI 層の import 規則を
.golangci.yml の1箇所で読めるようにする。遮断の理由も desc で
違反箇所に出るようになる。

internal 可視性で強制していると書いていた doc コメントを実態へ直す。
