+++ .golangci.yml
@@ -115,16 +115,15 @@ linters:
             - pkg: "github.com/hajimehoshi/ebiten$"
               desc: ebiten/v2を使う
         # レイアウトエンジンの接続点は UI プリミティブ実体の1箇所に限る。画面はもちろん、
-        # widgets の部品も直接は使わない。プリミティブ実体は internal 可視性で画面から
-        # 遮断済みだが、furex は通常の依存なので import 規則で封じ込める
+        # widgets の部品も直接は使わない
         layout_engine_guard:
           list-mode: lax
           files:
             - "$all"
-            - "!**/internal/widgets/internal/uicore/**"
+            - "!**/internal/widgets/uicore/**"
           deny:
             - pkg: "github.com/yohamta/furex"
-              desc: レイアウトエンジンは widgets/internal/uicore だけが使う。部品と画面は Container・FlexColumn を使う
+              desc: レイアウトエンジンは widgets/uicore だけが使う。部品と画面は Container・FlexColumn を使う
         # 部品は自分を使う上位層を知らない。widgets から 画面 states・画面制御 menuloop・
         # 組み立て maingame への逆依存を禁じ、上位が部品を使う一方向を保つ。Go は import 循環を
         # 弾くが循環しない一方向の逆依存は許すため、import 規則で塞ぐ。統合検証のテストは除外する
@@ -167,17 +166,28 @@ linters:
           allow:
             - "$gostd"
             - "github.com/hajimehoshi/ebiten/v2"
-            - "github.com/kijimaD/ruins/internal/widgets/internal/uicore"
+            - "github.com/kijimaD/ruins/internal/widgets/uicore"
         ui_core_guard:
           list-mode: strict
           files:
-            - "**/internal/widgets/internal/uicore/**"
+            - "**/internal/widgets/uicore/**"
             - "!$test"
           allow:
             - "$gostd"
             - "github.com/hajimehoshi/ebiten/v2"
             - "github.com/yohamta/furex"
             - "github.com/go-text/typesetting"
+        # プリミティブ実体を直接使えるのは widgets の内側だけにする。画面 states と画面制御
+        # menuloop は再輸出だけを持つファサード widgets/ui を通す。組み立てたツリーを描く
+        # 統合検証のテストも widgets の外には置かないので、テストの除外は要らない
+        ui_core_inbound_guard:
+          list-mode: lax
+          files:
+            - "$all"
+            - "!**/internal/widgets/**"
+          deny:
+            - pkg: "github.com/kijimaD/ruins/internal/widgets/uicore"
+              desc: プリミティブ実体は widgets の部品だけが使う。画面はファサード widgets/ui を使う
         # キー読み取り能力への import 許可制。読み取り API を持つパッケージを
         # 使える場所を列挙し、それ以外からは新旧問わず全シンボルを塞ぐ。
         # 入力を束縛表へ集約し、デバイス層とごく少数の例外だけが直接読む。テストは
