+++ internal/widgets/doc.go
@@ -34,7 +34,7 @@
 //   - ビジネスロジックとの連携が必要
 //   - 単体テストを書きたい
 //
-// ## internal/uicore を直接使う場合
+// ## widgets/uicore を直接使う場合
 //   - 保持型のツリーで画面を宣言的に組みたい
 //   - 状態管理は state 側が持ち、描画部品だけが要る
 //   - Canvas を差し替えてテストしたい
