+++ internal/widgets/entityspec/uipanel_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/kijimaD/ruins/internal/widgets/entityspec"
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 )
 
 // BuildSpecPanel の検証はツリー構造で行う。CollectLabels は描画せず Value を集めるだけなので
