+++ internal/widgets/hud/infopanel.go
@@ -5,8 +5,8 @@ import (
 
 	text "github.com/hajimehoshi/ebiten/v2/text/v2"
 
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
 	"github.com/kijimaD/ruins/internal/widgets/theme"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 )
 
 // 画面右上へ置く情報パネルの寸法
