+++ internal/widgets/hud/status_badges.go
@@ -7,8 +7,8 @@ import (
 	"slices"
 
 	"github.com/hajimehoshi/ebiten/v2/text/v2"
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
 	theme "github.com/kijimaD/ruins/internal/widgets/theme"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 )
 
 // StatusBadge はステータスバッジの情報
