+++ internal/widgets/menuframe/list_test.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
 	"github.com/kijimaD/ruins/internal/widgets/menuframe"
 	"github.com/kijimaD/ruins/internal/widgets/styled"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 )
 
 // menuframe.RenderList の検証はツリー構造で行う。CollectLabels は描画せず Value を集めるだけなので
