+++ internal/widgets/messagelog/golden_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/kijimaD/ruins/internal/gamelog"
 	"github.com/kijimaD/ruins/internal/testutil"
 	"github.com/kijimaD/ruins/internal/vrt"
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
 	"github.com/kijimaD/ruins/internal/widgets/messagelog"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 
 	"github.com/kijimaD/ruins/internal/world/query"
 )
