+++ internal/widgets/overlay/detailui.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/kijimaD/ruins/internal/resources"
 	"github.com/kijimaD/ruins/internal/widgets/entityspec"
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
 	"github.com/kijimaD/ruins/internal/widgets/pagination"
 	"github.com/kijimaD/ruins/internal/widgets/theme"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 )
 
 // buildDetailUI は性能行の並びから詳細モーダルを internal/uicore のツリーとして組み、rect いっぱいに配置して返す。
