+++ internal/widgets/overlay/overlay.go
@@ -3,7 +3,7 @@ package overlay
 import (
 	"image"
 
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 	w "github.com/kijimaD/ruins/internal/world"
 )
