+++ internal/widgets/uicore/golden_test.go
@@ -7,10 +7,10 @@ import (
 	"github.com/hajimehoshi/ebiten/v2"
 
 	"github.com/kijimaD/ruins/internal/vrt"
-	"github.com/kijimaD/ruins/internal/widgets/internal/uicore"
+	"github.com/kijimaD/ruins/internal/widgets/uicore"
 )
 
-// TestGolden_UISpecPanel は internal/uicore で組んだスペックパネルの実ピクセルを golden で固定する。
+// TestGolden_UISpecPanel は widgets/uicore で組んだスペックパネルの実ピクセルを golden で固定する。
 // EbitenCanvas で実フォント描画し、testdata/TestGolden_UISpecPanel.png と比較する。
 // フェイスはプールから借りて独立させるので、ロック無し・t.Parallel で回せる。
 func TestGolden_UISpecPanel(t *testing.T) {
