+++ internal/widgets/uicore/testdata/TestGolden_UISpecPanel.png
