+++ internal/widgets/uicore/text.go
